fix funcionalty

+++ src/main/java/com/example/productapi/repository/CategoryRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends ReactiveCrudRepository<Category, Long> {
+public interface CategoryRepository extends ReactiveCrudRepository<Category, Integer> {
 }
